# docs/02-engineering/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sergeant API",
    "version": "v1",
    "description": "Сгенеровано з zod-схем у `@sergeant/shared/schemas/api`. Не редагуй вручну — `pnpm api:generate-openapi` перегенерує."
  },
  "servers": [
    {
      "url": "https://sergeant.app",
      "description": "Production (Vercel edge → Coolify/Hetzner)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev (apps/server, default Express PORT)"
    }
  ],
  "paths": {
...
    "/api/push/register": {
      "post": {
        "summary": "Зареєструвати push-пристрій (web/ios/android)",
        "tags": [
          "push"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRegister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Bad request — payload не пройшов zod-валідацію.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — потрібна активна сесія.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/push/unregister": {
      "post": {
        "summary": "Зняти push-пристрій",
        "tags": [
          "push"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushUnregister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Bad request — payload не пройшов zod-валідацію.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — потрібна активна сесія.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/push/send": {
      "post": {
        "summary": "Internal-only fan-out push (worker → cron job)",
        "tags": [
          "push"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSend"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Send summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSendSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad request — payload не пройшов zod-валідацію.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PushRegister": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "platform": {
                "type": "string",
                "const": "web"
              },
              "token": {
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              },
              "keys": {
                "type": "object",
                "properties": {
                  "p256dh": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "auth": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  }
                },
                "required": [
                  "p256dh",
                  "auth"
                ]
              }
            },
            "required": [
              "platform",
              "token",
              "keys"
            ]
          },
          {
            "type": "object",
            "properties": {
              "platform": {
                "type": "string",
                "enum": [
                  "ios",
                  "android"
                ]
              },
              "token": {
                "type": "string",
                "minLength": 1,
                "maxLength": 4096
              }
            },
            "required": [
              "platform",
              "token"
            ]
          }
        ],
        "description": "POST /api/push/register — discriminated union web/ios/android.",
        "type": "object"
      },
      "PushUnregister": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "platform": {
                "type": "string",
                "const": "web"
              },
              "endpoint": {
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              }
            },
            "required": [
              "platform",
              "endpoint"
            ]
          },
          {
            "type": "object",
            "properties": {
              "platform": {
                "type": "string",
                "enum": [
                  "ios",
                  "android"
                ]
              },
              "token": {
                "type": "string",
                "minLength": 1,
                "maxLength": 4096
              }
            },
            "required": [
              "platform",
              "token"
            ]
          }
        ],
        "description": "POST /api/push/unregister.",
        "type": "object"
      },
...
      "PushSend": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "body": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          },
          "module": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ]
          },
          "tag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "userId",
          "title"
        ],
        "description": "Internal /api/push/send (worker)."
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "error",
          "message",
          "code"
        ],
        "additionalProperties": false,
        "description": "Канонічна shape body-помилки від `errorHandler` (4xx/5xx через AppError-ієрархію)."
      },
...
      "PushSendSummary": {
        "type": "object",
        "properties": {
          "delivered": {
            "type": "object",
            "properties": {
              "ios": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "android": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "web": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "ios",
              "android",
              "web"
            ],
            "additionalProperties": false
          },
          "cleaned": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "platform": {
                  "type": "string",
                  "enum": [
                    "ios",
                    "android",
                    "web"
                  ]
                },
                "reason": {
                  "type": "string"
                }
              },
              "required": [
                "platform",
                "reason"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "delivered",
          "cleaned",
          "errors"
        ],
        "additionalProperties": false,
        "description": "Уніфікований summary push fan-out (delivered/cleaned/errors)."
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Better Auth session cookie. Web-клієнт надсилає автоматично; mobile використовує bearerAuth."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Mobile (Expo) bearer-токен з Better Auth. Web використовує cookieAuth."
      }
    }
  }
}